mcp-project/agents/python-bridge/requirements.txt
# Core dependencies
nats-py==2.3.0
pydantic==2.5.0
loguru==0.7.1
fastapi==0.110.0
uvicorn==0.24.0
requests==2.31.0

# AI and ML dependencies
smolagents==1.13.0.dev0
transformers==4.40.2
torch==2.3.0
sentencepiece==0.1.99

# Testing dependencies
pytest==8.1.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0

# Development tools
black==23.7.0
isort==5.12.0
mypy==1.5.1

# Optional: Telemetry support
# Uncomment these to enable OpenTelemetry monitoring
# opentelemetry-sdk==1.20.0
# opentelemetry-exporter-otlp==1.20.0
# openinference-instrumentation-smolagents==0.1.0